+++ app/src/main/java/com/kojingu/xylo/standbyscreen/IconStandby.kt
@@ -1,0 +1,22 @@
+package com.kojingu.xylo.standbyscreen
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun IconStandby(
+    modifier: Modifier = Modifier
+) {
+    CircularProgressIndicator(
+        modifier = modifier
+            .size(60.dp)
+            .padding(top = 40.dp),
+        color = Color.Black,
+        strokeWidth = 6.dp
+    )
+}
